# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dokandar Mama shop assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: Shops the signed-in user belongs to
      responses:
        "200":
          description: Shops
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ShopWithRole" }
    post:
      operationId: createShop
      tags: [shops]
      summary: Create the user's shop (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShopInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shop" }
  /shops/current:
    get:
      operationId: getCurrentShop
      tags: [shops]
      summary: Active shop, role and organization
      responses:
        "200":
          description: Current shop
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentShop" }
    patch:
      operationId: updateCurrentShop
      tags: [shops]
      summary: Update shop settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShopUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shop" }
  /shops/branches:
    post:
      operationId: createBranch
      tags: [shops]
      summary: Add a branch to the organization (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShopInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shop" }
  /shops/current/members:
    get:
      operationId: listShopMembers
      tags: [shops]
      summary: Team members of the active shop
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ShopMember" }
    post:
      operationId: addShopMember
      tags: [shops]
      summary: Add or update a team member (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShopMemberInput" }
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShopMember" }
  /shops/current/members/{id}:
    delete:
      operationId: removeShopMember
      tags: [shops]
      summary: Remove a team member (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed
...
components:
  schemas:
    Shop:
      type: object
      required: [id, name, category, subscriptionPlan, createdAt]
      properties:
        id: { type: integer }
        organizationId: { type: ["integer", "null"] }
        name: { type: string }
        category: { type: string }
        ownerName: { type: ["string", "null"] }
        area: { type: ["string", "null"] }
        subscriptionPlan: { type: string, enum: [basic, standard, premium, organization] }
        cashboxAddon: { type: boolean }
        createdAt: { type: string }
    ShopWithRole:
      allOf:
        - $ref: "#/components/schemas/Shop"
        - type: object
          required: [role]
          properties:
            role: { type: string, enum: [admin, manager, shopkeeper] }
    ShopInput:
      type: object
      required: [name, category]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        ownerName: { type: string }
        area: { type: string }
        organizationName: { type: string }
    ShopUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        ownerName: { type: string }
        area: { type: string }
        subscriptionPlan: { type: string, enum: [basic, standard, premium, organization] }
        cashboxAddon: { type: boolean }
    CurrentShop:
      type: object
      properties:
        shop:
          oneOf:
            - $ref: "#/components/schemas/Shop"
            - type: "null"
        role: { type: ["string", "null"], enum: [admin, manager, shopkeeper, null] }
        organization:
          oneOf:
            - $ref: "#/components/schemas/OrganizationRef"
            - type: "null"
    OrganizationRef:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
    ShopMember:
      type: object
      required: [id, userId, role]
      properties:
        id: { type: integer }
        userId: { type: string }
        name: { type: ["string", "null"] }
        role: { type: string, enum: [admin, manager, shopkeeper] }
        createdAt: { type: string }
    ShopMemberInput:
      type: object
      required: [userId, role]
      properties:
        userId: { type: string, minLength: 1 }
        name: { type: string }
        role: { type: string, enum: [admin, manager, shopkeeper] }
...