# lib/api-zod/src/generated/types/shopUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Dokandar Mama shop assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ShopUpdateSubscriptionPlan } from './shopUpdateSubscriptionPlan';

export interface ShopUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  ownerName?: string;
  area?: string;
  subscriptionPlan?: ShopUpdateSubscriptionPlan;
  cashboxAddon?: boolean;
}